Remove specific perl version
Actually 5.008 is the minimun version I would ever allow my friends to use
after they convinced me that upgrading to 5.14.x was just something they
could not do.

But the code doesn't have, as far as I know, anything that would prevent
it from working on something older, so I'm removing this restriction.

Signed-off-by: Pedro Melo <melo@simplicidade.org>